perf: remove temporary allocations in Parquet
This removes more than 50% of the temporary allocations in the Parquet readers.
The only temporary allocations left are in the Thrift definitions and the
`polars_io` code.
